Extract converge into own file
