+++ packages/interactor/src/index.ts
@@ -1,8 +1,4 @@
-import { main, timeout as effectionTimeout } from 'effection';
-
-// TODO: this API is available on browsers as `window.performance`, we need to figure out
-// a way to package this so it'll work on both browsers and node.
-import { performance } from 'perf_hooks';
+import { converge } from './converge';
 
 interface Options {
   timeout: number,
@@ -13,23 +9,6 @@ let defaultOptions: Options = {
   timeout: 1900
 }
 
-async function converge<T>(timeout: number, fn: () => T): Promise<T> {
-  return await main(function*() {
-    let startTime = performance.now();
-    while(true) {
-      try {
-        return fn();
-      } catch(e) {
-        let diff = performance.now() - startTime;
-        if(diff > timeout) {
-          throw e;
-        } else {
-          yield effectionTimeout(1);
-        }
-      }
-    }
-  });
-}
 
 export interface InteractorSpecification {
   name: string,
